Fixed Water Wading SFX fade-out
Added a collider where the cave water is which triggers a new state, called InWater. InWater provides more precise control to determine whether 'FS_Water_Wade' should be muted or not. Through this implementation, the tails of water wading remain, but their improper continuation when the player exits the water is now hasta la bye-bye.